requirements.txt
fastapi==0.110.0
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pillow==10.1.0

pydantic>=2.12.5,<3.0.0
pydantic-settings>=2.6.1,<3.0.0
fastapi-mail==1.6.1